pyproject.toml
[project]
name = "ocr-verification"
version = "0.1.0"
description = "FastAPI server for document processing with DSPy and OCR"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy-ai>=3.0.4",
    "easyocr>=1.7.2",
    "opencv-python>=4.8.0", # Required by easyocr for cv2
    "langgraph>=1.0.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12", # For file uploads
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0", # For settings management
    "python-json-logger>=3.2.1", # Structured logging
    "pdf2image>=1.17.0",
    "pillow>=10.0.0", # For image processing
    "paddlepaddle>=3.2.2",
    "paddleocr>=3.3.2",
    "paddlex[ocr]>=3.3.10", # Required for PaddleOCR-VL
    "google-cloud-vision>=3.8.0", # Google Cloud Vision API for cloud OCR
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
